Make more concise and for easier future editing
The `PS1` method is another legitimate way to test for interactivity, -
while saving a few characters. lol What can I say, I'm bored, and
looking for ways to improve code in any way of which I can think.

Added a `return 0` to the end of the file, just in-case an error
occurs, causing some confusing exit statuses.

Fun fact: the `source` command can also be written as `.` (period).

Cleaned up the `PATH` check-and-amend procedure.