Optimize and prefer `sh`
Since there's nothing `bash`-related here, I've opted for the smaller
Bourne Shell. As for `grep`, it was redundant as AWK does that without
breaking a sweat.

However, check this works first. If I had a copy of the output you're
parsing, I could do some good stuff, but as it is, my hands are tied.