Prefer `grep -E` over deprecated `egrep`
I've also improved the `apt-get` usage in one of your aliases, as
repeated in similar files. The reason is that, if you look at the
`apt-get` manual page, you'll see `clean` already implies what
`autoclean` does, but it removes it all, so using both is redundant.

Your `egrep` and `fgrep` aliases should work the same, but will
continue to do so when the deprecation period is over and `egrep` then
vanishes in favor of `grep -E`. Same with `grep -F` and `fgrep`, I
presume.